requirements.txt
# =============================================================================
# EC2 — ETL Loader + API Server
# Instalar con: pip install -r requirements.txt
# =============================================================================

# AWS S3
boto3>=1.28.0

# PostgreSQL
psycopg2-binary>=2.9.0

# API Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Utilidades
python-dotenv>=1.0.0
requests>=2.31.0
